// src/types/Usuarios.ts
type User = {
  _id: string;
  nombre: string;
    apellido: string;
    telefono?: string;
    nombre_de_usuario: string;
    credencial?: string;
    unidad: string;
    jerarquia: string;
    plaza?: string;
    zona: string;
    rol: string;
    imagen: string;
    createdAt: string;
    updatedAt: string;
    __v: number;
    usuario_repo?: string;


}

export default User;